chore: move clickhousereader filter suggestion methods to dedicated file (#6061)
